Supermarket code changes
The above attached code is for Supermarket items total price calculation, which calculates as per the problem statement mentioned. It also considers the offers and their special price. It calculated price as items are added. To execute please run class PayNowForShopping.java and pass items you like to purchase, for e.g "java PayNowForShopping A A A B B C D A". Then it will calculate the total price of the items passed here.